All player input types work without modes
The mode system predefined which input device the player was aloud to use
at one time. Now, the input check goes through each if the player has not
entered an input on another device. i.e. if the motion axis is still 0
after the keyboard check, it will check the joystick, then the d-pad, etc.