Fixed drawing at origin issue
It turns out that transform origins weren't getting passed into sprite
draw. That's why I needed that GetWorldPosition function for a while.
Deleted it. Also deleted shear from entity and draw, because for
whatever reason it was causing the sprite to not draw at all. I didn't
really need it anyway.